Fix errors that surfaced by adding /sdl compiler flag to the build

+++ lib/Bitcode/Reader/BitcodeReader.cpp
@@ -2393,7 +2393,7 @@ uint64_t BitcodeReader::decodeSignRotatedValue(uint64_t V) {
   if ((V & 1) == 0)
     return V >> 1;
   if (V != 1)
-    return -(V >> 1);
+    return -(int64_t)(V >> 1);
   // There is no such thing as -0 with integers.  "-0" really means MININT.
   return 1ULL << 63;
 }
